+++ src/main/java/kea/alog/aggregator/service/openfeign/project/ProjectFeign.java
@@ -1,11 +1,10 @@
-package kea.alog.aggregator.service.openfeign;
+package kea.alog.aggregator.service.openfeign.project;
 
 import kea.alog.aggregator.common.dto.PageDto;
 import kea.alog.aggregator.common.dto.ResponseDto;
 import kea.alog.aggregator.config.JwtRequestFeignConfiguration;
 import kea.alog.aggregator.web.constant.ProjectSortType;
 import kea.alog.aggregator.web.dto.ProjectDto.FeignProjectResponseDto;
-import kea.alog.aggregator.web.dto.TopicDto.FeignTopicResponseDto;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -28,7 +27,4 @@ public interface ProjectFeign {
     @GetMapping("{projectPk}/members")
     ResponseDto<PageDto<Long>> findMembers(@PathVariable("projectPk") Long projectPk, @RequestParam(value = "keyword", required = false) String keyword,
         @RequestParam("page") int page, @RequestParam("size") int size);
-
-    @GetMapping("{projectPk}/topics/{topicPk}")
-    ResponseDto<FeignTopicResponseDto> findTopic(@PathVariable("projectPk") Long projectPk, @PathVariable("topicPk") Long topicPk);
 }
